New commands
Added a /tp command, allows players to teleport to eachother, also should allow for any vehicles the player is in to be teleported, along with passangers. Needs testing as I could not locate enough players to test functionality properly.

Added a /flip command to turn cars that have flipped over.

Few memory improvements